✏️ WV-12 rename : 폴더 이름 변경
